Extract truncate_tests.rs from turn.rs (parent #680)
Move the remaining #[cfg(test)] mod truncate_tests originally embedded
near the bottom of turn.rs (~620 LOC) into a new
src/agent/loop_run/truncate_tests.rs sibling file. This is the final
test-mod extraction; turn.rs now contains zero #[cfg(test)] mod blocks.

Strategy (same as turn_tests / progress_tests):
- Move the entire mod verbatim.
- Rewrite the inner mod's `use super::{...}` block to source each item
  from its actual sibling module (7 items across 6 modules).
- Add file-level explicit `use super::verifier_orchestration::{...}` for
  4 names accessed via qualified `super::X` in test bodies.
- Rename inner `mod truncate_tests` → `mod inner` (clippy::module_inception).
- `cargo fix --tests` pruned unused entries from the mirrored prelude.

#[cfg(test)] only; production binary excludes this file. No facade
re-export (DR3-001).

turn.rs: 7,806 -> 7,179 (-627 LOC)
Cumulative from 39,489: -32,310 LOC (-81.8%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
